Fixes to spot configs
Organizing spot member vars to be all self (no global to class)

Fixing up script for setting spots

Correcting log info to try to address #1

Changing which dir gets spots logs, so maybe fixing #2 as well

+++ src/utils/Cam.py
@@ -64,7 +64,7 @@ class Cam:
             Spots[sn].imageTimeStamp = ts
             
             # Get the polygon vertices for the spot
-            verts = array( Spots[sn].vertices )
+            verts = array( Spots[sn].vertices ).astype('int32')
             
             # Make (boolean) mask for spot
             mask = zeros((im.shape[0],im.shape[1]))
